requirements.txt
netifaces>=0.11.0
python-nmap>=0.7.1
psutil>=5.9.0
paramiko>=2.9.0
pywinrm>=0.4.3